requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.21.0
pyyaml>=5.4.0
argparse>=1.1
spikingjelly>=0.0.0.0.14
timm>=0.6.0
thop>=0.1.1
tensorboard>=2.8.0
cupy-cuda12x>=12.0.0
fastapi
uvicorn[standard]
torch
torchvision
timm
mtcnn
pyyaml
pillow
numpy
tensorflow
python-multipart>=0.0.6
facenet-pytorch